requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.2
OpenAIAuth==3.0.0
prompt-toolkit==3.0.39
Pygments==2.16.1
PySocks==1.7.1
python-telegram-bot==20.2
regex==2023.8.8
requests==2.28.2
revChatGPT==6.8.6
rfc3986==1.5.0
rich==13.5.2
sniffio==1.3.0
socksio==1.0.0
tiktoken==0.4.0
tls-client==0.2.1
tqdm==4.65.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
